+++ CHANGELOG.md
@@ -202,6 +202,10 @@
 - Added settings access in Sodium.
 - Updated Simplified Chinese translations ([#242](https://github.com/LambdAurora/LambDynamicLights/pull/242)).
 
+### 3.1.1
+
+- Fixed in-world item light source data reload not applying.
+
 ## 3.2.0
 
 - Updated to Minecraft 1.21.2.
@@ -210,6 +214,11 @@
 - Tweaked the new dynamic lighting engine introduced in 3.1.0.
 - Updated Estonian translations ([#243](https://github.com/LambdAurora/LambDynamicLights/pull/243)).
 - Updated [SpruceUI].
+
+### 3.2.1
+
+- Same changes as v3.1.1 but for 1.21.3.
+- Fixed in-world item light source data reload not applying.
 
 [SpruceUI]: https://github.com/LambdAurora/SpruceUI "SpruceUI page"
 [pridelib]: https://github.com/Queerbric/pridelib "Pridelib page"